pyproject.toml
[project]
name = "stock-portfolio-ai"
version = "0.1.0"
description = "Multi-agent stock research and portfolio management system."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain",
    "langchain-openai",
    "langgraph",
    "pandas",
    "pydantic",
    "python-dotenv",
    "yfinance",
]

[project.scripts]
stock-portfolio-ai = "stock_portfolio_ai:main"

[build-system]
requires = ["uv_build>=0.11.7,<0.12.0"]
build-backend = "uv_build"
